Continue scale-down when certain exceptions are encountered
All of these scenarios don't really require or benefit from stopping the
autoscaler run; stopping on any of these simply slows down the scaling
process.

This is ESPECIALLY true when there are multiple of these at once. It
takes an autoscaler run to resolve EACH issue (when it can be safely
done all at once), and when this is run by a Lambda on a schedule, this
means scale-ups can take many minutes to actually add a worker (the
most-frequent Lambda schedule is 1-minute intervals).