Improve scaling efficiency when there are cleanup tasks or errors (#38)
* Continue scale-down when certain exceptions are encountered

All of these scenarios don't really require or benefit from stopping the
autoscaler run; stopping on any of these simply slows down the scaling
process.

This is ESPECIALLY true when there are multiple of these at once. It
takes an autoscaler run to resolve EACH issue (when it can be safely
done all at once), and when this is run by a Lambda on a schedule, this
means scale-ups can take many minutes to actually add a worker (the
most-frequent Lambda schedule is 1-minute intervals).

* Add decision comments to log output

* Include scale up/down numbers in decision comments

* Clean up all stray instances at once

There's not really any good reason to only clean up one stray at at
time. If it's a stray, it's a stray; just clean them all up.

* Add instance count in log output

* Tweak "no decision" message for clarity

* Drop continue labels